refactor: ♻️ separate consts, use barrel pattern
